feat: implements manage of foundations and donation partially

+++ client/src/app/foundations/manage/manage.component.ts
@@ -9,7 +9,7 @@ import { Account } from 'src/app/auth/auth.dto';
 import { FoundationUpdateDTO } from '@models/foundation';
 import { MatAutocompleteSelectedEvent, MatDialog, MatTable } from '@angular/material';
 import { Product } from '@models/product';
-import { ManageProductComponent } from '@dialogs/index';
+import { ManageProductComponent, ConfirmDialogComponent, ConfirmData } from '@dialogs/index';
 import { Item, ItemUnit } from '@models/fields/item';
 import { FormControl } from '@angular/forms';
 import { ProductService } from '@core/product.service';
@@ -18,6 +18,7 @@ import { Collaborator, Role, RoleValues } from '@models/fields/collaborator';
 import { ActionCategory, ActionCategoryValues } from '@models/campaign';
 import { ItemUnitValues } from '../../shared/models/fields/item';
 import { UsersService } from 'src/app/users/users.service';
+import { AddressService } from '@core/address.service';
 
 enum FoundationTab {
   REGISTER_DATA,
@@ -32,6 +33,7 @@ enum FoundationTab {
 })
 export class ManageComponent implements OnInit, OnDestroy {
 
+  private cep: Subject<void>;
   private _destroy$: Subject<void>;
 
   private userColumns: string[];
@@ -43,6 +45,7 @@ export class ManageComponent implements OnInit, OnDestroy {
   private userTable: MatTable<Item>;
 
   private isLoading: boolean;
+  private isLoadingAddress: boolean;
 
   private logged: Account;
   private selectedTab: FoundationTab;
@@ -50,10 +53,10 @@ export class ManageComponent implements OnInit, OnDestroy {
   private foundation: FoundationUpdateDTO;
   private categoryOptions: ActionCategory[];
 
-  private defaultManager: Collaborator;
-  private collaborator: Collaborator;
   private userRoles: Role[];
   private userControl: FormControl;
+  private collaborator: Collaborator;
+  private defaultManager: Collaborator;
   private collaborators: Collaborator[];
   private users: Observable<Collaborator[]>;
 
@@ -70,20 +73,23 @@ export class ManageComponent implements OnInit, OnDestroy {
     private readonly _userService: UsersService,
     private readonly _productService: ProductService,
     private readonly _accountService: AccountService,
+    private readonly _addressService: AddressService,
     private readonly _foundationService: FoundationsService
   ) {
-    this.userColumns = ['name', 'email', 'role', 'actions'];
+    this.userColumns = ['avatar', 'name', 'email', 'role', 'actions'];
     this.productColumns = ['product', 'quantity', 'unit', 'actions'];
     this.selectedTab = FoundationTab.REGISTER_DATA;
 
+    this.cep = new Subject();
     this._destroy$ = new Subject();
 
     this.foundation = new FoundationUpdateDTO();
     this.categoryOptions = ActionCategoryValues;
 
-    this.collaborator = new Collaborator();
     this.userRoles = RoleValues;
     this.userControl = new FormControl();
+    this.collaborator = new Collaborator();
+    this.defaultManager = new Collaborator();
 
     this.item = new Item();
     this.goal = new GoalCreateDTO();
@@ -140,28 +146,69 @@ export class ManageComponent implements OnInit, OnDestroy {
         distinctUntilChanged(),
         map(value => typeof value === 'string' ? value : value.name),
         switchMap(value => this._userService.listCollaborators(value)),
+        map(response => {
+          if (!response.length) {
+            return [{
+              name: 'Não encontrado',
+              email: 'Nenhum resultado para esta pesquisa'
+            }] as Collaborator[];
+          }
+          return response;
+        })
       );
+    this.cep.pipe(
+      takeUntil(this._destroy$),
+      debounceTime(300),
+    )
+      .subscribe(async () => {
+        const { cep } = this.foundation.address;
+        if (cep.length === 8) {
+          try {
+            this.isLoadingAddress = true;
+            this.foundation.address = await this._addressService.getByCEP(cep);
+          } catch (e) {
+            console.error(e);
+          } finally {
+            this.isLoadingAddress = false;
+          }
+        }
+      });
   }
 
   ngOnDestroy() {
     this._destroy$.next();
   }
 
+  async onUpdateFoundation() {
+    try {
+      this.isLoading = true;
+      this.foundation = await this._foundationService.update(this.foundation);
+    } catch (e) {
+      console.error(e);
+    } finally {
+      this.isLoading = false;
+    }
+  }
+
   /** Product functions */
+  onResetItem() {
+    this.productControl.setValue('');
+    this.item = new Item();
+  }
+
   productLabel(payload?: Product): string | undefined {
     return payload && payload._id && payload.name;
   }
 
-  onRemoveItem(value: Product) {
-    const index = this.goal.items.findIndex(item => value._id === item.product);
+  onRemoveItem(value: Item) {
+    const index = this.goal.items.findIndex(item => item.product === value.product);
     this.goal.items.splice(index, 1);
     this.goalTable.renderRows();
   }
 
   onInsertItem() {
     this.goal.items.push({ ...this.item });
-    this.item = new Item();
-    this.productControl.setValue('');
+    this.onResetItem();
     this.goalTable.renderRows();
   }
 
@@ -185,12 +232,29 @@ export class ManageComponent implements OnInit, OnDestroy {
   }
 
   /** User functions */
+  onResetCollaborator() {
+    this.userControl.setValue('');
+    this.collaborator = new Collaborator();
+  }
+  
   userLabel(payload?: Collaborator): string | undefined {
     return payload && payload.user && payload.name;
   }
 
   async onRemoveCollaborator(value: Collaborator) {
     try {
+      const result = await this._dialog.open(ConfirmDialogComponent, {
+        data: new ConfirmData(
+          'Remover colaborador',
+          'Você realmente deseja remover este colaborador?',
+          'Não',
+          'Sim',
+          true
+        )
+      })
+        .afterClosed()
+        .toPromise();
+      if (!result) { return; }
       await this._foundationService.removeCollaborator(value);
       const index = this.collaborators.findIndex(col => value.user === col.user);
       this.collaborators.splice(index, 1);
@@ -209,12 +273,16 @@ export class ManageComponent implements OnInit, OnDestroy {
       } else {
         this.collaborators.push({ ...this.collaborator });
       }
-      this.collaborator = new Collaborator();
-      this.userControl.setValue('');
+      this.onResetCollaborator();
       this.userTable.renderRows();
     } catch (e) {
       console.error(e);
     }
+  }
+
+  onEditCollaborator(value: Collaborator) {
+    this.userControl.setValue(value);
+    this.collaborator = { ...value };
   }
 
   async onSelectUser(evt: MatAutocompleteSelectedEvent) {
@@ -224,12 +292,30 @@ export class ManageComponent implements OnInit, OnDestroy {
     }
   }
 
+  /** Goal functions */
+  private async onSaveGoal() {
+    try {
+      this.isLoading = true;
+      this.goal = await this._foundationService.saveGoal(this.goal);
+      this.goalTable.renderRows();
+    } catch (e) {
+      console.error(e);
+    } finally {
+      this.isLoading = false;
+    }
+  }
+
   private async fetchFoundation() {
     try {
       this.isLoading = true;
+      const goal = await this._foundationService.getGoal();
+      if (goal) {
+        this.goal = goal;
+      }
       this.foundation = await this._foundationService.get(this.logged._id);
-      this.collaborators = await this._foundationService.listCollaborators();
-      this.defaultManager = this.collaborators.find(col => col.email === this.logged.email);
+      const collaborators = await this._foundationService.listCollaborators();
+      this.defaultManager = collaborators.find(col => col.email === this.logged.email);
+      this.collaborators = collaborators.filter(col => col.email !== this.defaultManager.email);
     } catch (e) {
       console.error(e);
     } finally {
